Add files via upload
This is essentially a blank start to the FYP version 1 STM code for STM version 1.16.1. All I have done is add pins to the IOC corresponding to those PCB version 1 connects to and confirmed that the IDE generates code. If code is not generating, try logging into my ST and downloading the latest NUCLEO F0 firmware.